requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.56.1
django-environ==0.10.0
Faker==19.12.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.5
